backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
numpy==1.24.3
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
diffusers==0.21.4
transformers==4.31.0
accelerate==0.21.0
huggingface-hub==0.16.4
python-engineio==4.8.2
python-socketio==5.10.0 